integrate product service with data api, file api, and mvc layer
complete product flow with base service, file upload, and mvc integration

+++ App/App.Models.DTO/Product/ProductDetailDto.cs
@@ -6,7 +6,7 @@
         public string Name { get; set; } = null!;
         public string Details { get; set; } = null!;
         public decimal Price { get; set; }
-        public int StockAmount { get; set; }
+        public byte StockAmount { get; set; }
         public string CategoryName { get; set; } = null!;
         public List<string> Images { get; set; } = new();
         public int CategoryId { get; set; }
